examples/suit: Refactor to draft ietf-v3
