+++ examples/suit_update/coap_handler.c
@@ -11,7 +11,7 @@
 #include <string.h>
 
 #include "net/nanocoap.h"
-#include "suit/coap.h"
+#include "suit/transport/coap.h"
 
 static ssize_t _riot_board_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len, void *context)
 {
